+++ Assets/jwellone/PrefsEditor/Editor/AddEntityPopupWindowContent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6054cc1d907004333a34f223c01a690c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
