+++ Assets/jwellone/PrefsEditor/Editor/ConstGenerator.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d2a5ce9185e6f4213b409134f162ac5f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
