+++ Assets/jwellone/PrefsEditor/Editor/ConstGenerator/IPrefsConstGenerator.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+
+#nullable enable
+
+namespace jwelloneEditor
+{
+    public interface IPrefsConstGenerator
+    {
+        void Generate(IReadOnlyList<PrefsEntity> entities);
+    }
+}
