+++ Assets/jwellone/PrefsEditor/Editor/ConstGenerator/IPrefsConstGenerator.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 66e205062435447818d1fb9b4f845458
+guid: a8ca237ccc8f64763821aa615c1056f2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
