+++ Assets/jwellone/PrefsEditor/Editor/Entity.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 398897431b27444d78cf4292336fe962
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
