+++ Assets/jwellone/PrefsEditor/Editor/Provider.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: de5b76cab84e142e4904f13f835dd5f7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
