+++ Assets/jwellone/PrefsEditor/Editor/Provider/Mac.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ea553e0702fca468e85bc80e28f3dac9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
