+++ Assets/jwellone/PrefsEditor/Editor/Provider/Windows.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 135a9dbed691e483295dc6096fff0f44
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
